Added avatar to topbar and linked to external api. Added remember me functionality

+++ src/app/birthdays-module/topbar/topbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Router } from '@angular/router';
 import { Subject } from 'rxjs';
+import { AuthService } from 'src/app/services/auth.service';
 
 @Component({
   selector: 'app-topbar',
@@ -10,12 +11,18 @@ import { Subject } from 'rxjs';
 export class TopbarComponent {
   @Output() addFriendEvent = new EventEmitter();
   @Output() editFriendEvent = new EventEmitter();
+  imageUrl: string = "";
 
-  constructor(private router: Router) { }
+  constructor(private router: Router, private authService: AuthService) { }
+
+  ngOnInit() {
+    this.authService.getAvatar().subscribe((response: any) => this.imageUrl = response.image);
+  }
+
   onAddButtonClick() {
     this.addFriendEvent.emit();
   }
-  onEditButtonClick(){
+  onEditButtonClick() {
     console.log("received event");
     this.editFriendEvent.emit();
   }
@@ -23,6 +30,11 @@ export class TopbarComponent {
 
 
   logout() {
+    var user = JSON.parse(sessionStorage.getItem("loggedInUserEmail") || '');
+    if (user != '' && user.rememberMe == true) {
+      this.router.navigateByUrl('/');
+      return;
+    }
     sessionStorage.clear();
     this.router.navigateByUrl('/')
   }
